First working version with generics!
