+++ src/main.rs
@@ -1,9 +1,63 @@
 use std::io;
 mod redux;
-use redux::{ Store, Action, State, Todo, VisibilityFilter };
-use redux::Action::*;
-use redux::TodoAction::*;
-use redux::VisibilityFilter::*;
+use redux::{ Store };
+use Action::*;
+use TodoAction::*;
+use VisibilityFilter::*;
+
+#[derive(Clone, Debug)]
+pub struct State {
+    pub todos: Vec<Todo>,
+    pub visibility_filter: VisibilityFilter
+}
+
+impl State {
+    pub fn with_defaults() -> State {
+        State {
+            todos: Vec::new(),
+            visibility_filter: VisibilityFilter::ShowAll,
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct Todo {
+    pub id: i16,
+    pub title: String,
+    pub completed: bool,
+    pub deleted: bool,
+}
+
+impl Todo {
+    pub fn new(id: i16, title: String) -> Todo {
+        Todo {
+            id: id,
+            title: title,
+            completed: false,
+            deleted: false,
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+pub enum Action {
+    Todos(TodoAction),
+    Visibility(VisibilityFilter),
+}
+
+#[derive(Clone, Debug)]
+pub enum TodoAction {
+    Add(String),
+    Toggle(i16),
+    Remove(i16),
+}
+
+#[derive(Clone, Debug)]
+pub enum VisibilityFilter {
+    ShowActive,
+    ShowAll,
+    ShowCompleted,
+}
 
 fn reducer(state: State, action: Action) -> State {
     // Always return a new state
@@ -85,7 +139,7 @@ fn render(state: State) {
 }
 
 fn main() {
-    let mut store = Store::create_store(reducer);
+    let mut store = Store::create_store(reducer, State::with_defaults());
     store.subscribe(render);
 
     print_instructions();
